created a utility function for get findOptions for getPharmacists and added postingPlace into query
